requirements.txt
# for base
pyocr
pyautogui
keyboard
pytesseract
pillow

# for mathine translation
deepl

# for summarization
pysummarization
mecab-python3
unidic-lite
accel-brain-base